Added ownsAdjacentBuildingToHex method
This method checks whether a given user occupies any vertices of a
hexTile. Super useful when deciding whether a user get to produce when
the dice is rolled and whether a robber can rob a player. Also fixed
the ModelTest for canRobPlayer which had a few bugs in it (we weren’t
checking that the victim was adjacent to the robber)